requirements.txt
nltk~=3.9.1
spacy~=3.7.5
matplotlib~=3.9.2
setuptools~=75.1.0
graphviz~=0.20.3
pandas~=2.2.3
requests~=2.32.3
scikit-learn~=1.5.2
python-dotenv~=1.0.1
natsort~=8.4.0
torch~=2.7.0+cu126
datasets~=3.0.1
Send2Trash~=1.8.3
transformers~=4.43.0
numpy~=1.26.4
pytest~=8.3.3
pytest-cov~=6.0.0
bs4~=0.0.2
dask~=2024.12.1
bertopic~=0.16.4
joblib~=1.5.0
bitsandbytes~=0.45.5
river~=0.22.0